VCF pipeline script
Python script incorporating BASH commands using argparse and subprocess commands to generate a VCF. Outputs the number of variants, the average VCF QUAL score, and the VCF file name based on a read depth specified by the user. Script allows for variable FASTQ read files (1 or 2) but the current script uses and absolute path for the FASTA reference files. 
Also trying to combine the samtools processing commands into one command keeps piping to the next command but running into issues with the first samtools piped command. The script works but keeps outputting the help menu for samtools to the print screen.